fix: measure frontend times only when flag enabled
