+++ src/lib/features/frontend-api/frontend-api-service.ts
@@ -16,14 +16,12 @@ import {
 import { validateOrigins } from '../../util';
 import { BadDataError, InvalidTokenError } from '../../error';
 import {
-    FUNCTION_TIME,
     FRONTEND_API_REPOSITORY_CREATED,
     PROXY_REPOSITORY_CREATED,
 } from '../../metric-events';
 import { FrontendApiRepository } from './frontend-api-repository';
 import { GlobalFrontendApiCache } from './global-frontend-api-cache';
 import { ProxyRepository } from './proxy-repository';
-import metricsHelper from '../../util/metrics-helper';
 
 export type Config = Pick<
     IUnleashConfig,
@@ -63,8 +61,6 @@ export class FrontendApiService {
 
     private cachedFrontendSettings?: FrontendSettings;
 
-    private timer: Function;
-
     constructor(
         config: Config,
         stores: Stores,
@@ -76,19 +72,12 @@ export class FrontendApiService {
         this.stores = stores;
         this.services = services;
         this.globalFrontendApiCache = globalFrontendApiCache;
-
-        this.timer = (functionName) =>
-            metricsHelper.wrapTimer(config.eventBus, FUNCTION_TIME, {
-                className: 'FrontendApiService',
-                functionName,
-            });
     }
 
     async getFrontendApiFeatures(
         token: IApiUser,
         context: Context,
     ): Promise<FrontendApiFeatureSchema[]> {
-        const stopTimer = this.timer('getFrontendApiFeatures');
         const client = await this.clientForFrontendApiToken(token);
         const definitions = client.getFeatureToggleDefinitions() || [];
         const sessionId = context.sessionId || String(Math.random());
@@ -109,7 +98,6 @@ export class FrontendApiService {
                 }),
                 impressionData: Boolean(feature.impressionData),
             }));
-        stopTimer();
         return resultDefinitions;
     }
 
@@ -117,7 +105,6 @@ export class FrontendApiService {
         token: IApiUser,
         context: Context,
     ): Promise<FrontendApiFeatureSchema[]> {
-        const stopTimer = this.timer('getNewFrontendApiFeatures');
         const client = await this.newClientForFrontendApiToken(token);
         const definitions = client.getFeatureToggleDefinitions() || [];
         const sessionId = context.sessionId || String(Math.random());
@@ -139,7 +126,6 @@ export class FrontendApiService {
                 }),
                 impressionData: Boolean(feature.impressionData),
             }));
-        stopTimer();
         return resultDefinitions;
     }
